associate user id with imdb id, add a new table called user

+++ lumi_list/lib/database/favorite.dart
@@ -1,7 +1,7 @@
 /*Info of the favorites list associated with the ♥ button */
 import 'package:sqflite/sqflite.dart';
 import 'app_database.dart';
-
+import 'user.dart';
 class FavoriteDao {
   // add to favorites
   static Future<void> insertFavorite({
@@ -11,10 +11,14 @@ class FavoriteDao {
     int? rating,
     String? genre,
   }) async {
+    final userId = UserDao.getCurrentUserId();
+    if (userId == null) throw Exception("Please login first");
+
     final db = await AppDatabase.database;
     await db.insert(
       'favorites',
       {
+        'user_id': userId,
         'imdb_id': imdbId,
         'title': title,
         'poster': poster,
@@ -28,34 +32,44 @@ class FavoriteDao {
 
   // delete from favorites
   static Future<void> deleteFavorite(String imdbId) async {
+    final userId = UserDao.getCurrentUserId();
+    if (userId == null) throw Exception("Please login first");
     final db = await AppDatabase.database;
     await db.delete(
       'favorites',
-      where: 'imdb_id = ?',
-      whereArgs: [imdbId],
+      where: 'imdb_id = ? AND user_id = ?',
+      whereArgs: [imdbId, userId],
     );
   }
 
   // check if favorite
   static Future<bool> isFavorite(String imdbId) async {
+    final userId = UserDao.getCurrentUserId();
+    if (userId == null) throw Exception("Please login first");
     final db = await AppDatabase.database;
     final res = await db.query(
       'favorites',
-      where: 'imdb_id = ?',
-      whereArgs: [imdbId],
+      where: 'imdb_id = ? AND user_id = ?',
+      whereArgs: [imdbId, userId],
     );
     return res.isNotEmpty;
   }
 
   // read all favorites
-  static Future<List<Map<String, dynamic>>> getAllFavorites() async {
+  static Future<List<Map<String, dynamic>>> getAllFavorites([int? userId]) async {
     final db = await AppDatabase.database;
-    return db.query('favorites');
+    if (userId == null) {
+      return db.query('favorites');
+    } else {
+      return db.query('favorites', where: 'user_id = ?', whereArgs: [userId]);
+    }
   }
   
   static Future<String?> getMostFrequentGenre() async {
-
-    final favorites = await getAllFavorites();
+     final userId = UserDao.getCurrentUserId();
+    if (userId == null) throw Exception("Please login first");
+    
+    final favorites = await getAllFavorites(userId);
     if (favorites.isEmpty) return null;
     Map<String, int> genreCounts = {};
 
@@ -63,7 +77,6 @@ class FavoriteDao {
       String? genreStr = movie['genre']; 
       if (genreStr != null && genreStr != "No Info" && genreStr.isNotEmpty) {
         List<String> genres = genreStr.split(',').map((g) => g.trim()).toList();
-        
         for (var genre in genres) {
           genreCounts[genre] = (genreCounts[genre] ?? 0) + 1;
         }
